browser/components/feeds/test/chrome/test_maxSniffing.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=739040
-->
<head>
  <title>Test that we only sniff 512 bytes</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=739040">Mozilla Bug 739040</a>
<p id="display">
  <iframe id="validTestFrame" src="http://mochi.test:8888/tests/browser/components/feeds/test/valid-feed.xml"></iframe>
  <iframe id="unsniffableTestFrame" src="http://mochi.test:8888/tests/browser/components/feeds/test/valid-unsniffable-feed.xml"></iframe>
</p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 739040 **/
SimpleTest.waitForExplicitFinish();

addLoadEvent(function() {
  is($("validTestFrame").contentDocument.documentElement.id, "feedHandler",
     "valid feed should be sniffed");
  isnot($("unsniffableTestFrame").contentDocument.documentElement.id, "feedHandler",
        "unsniffable feed should not be sniffed");
});
addLoadEvent(SimpleTest.finish);

</script>
</pre>
</body>
</html>
